requirements.txt
flask-rest
Flask-Session
azure-functions
python-dotenv==1.0.0
semantic-kernel==0.5.1.dev0
azure-core
azure.search.documents==11.4.0b8
colorlog
json2html
tiktoken